# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Describe It - Spanish Learning API
  version: 1.0.0
  description: |
    A comprehensive API for visual-based Spanish learning through AI-powered image descriptions, 
    Q&A generation, vocabulary extraction, and translation services.
    
    This API enables users to:
    - Generate multi-style image descriptions in Spanish and English
    - Create interactive Q&A sessions for comprehension testing
    - Extract and translate vocabulary from descriptions
    - Track learning progress and export data
    - Search and manage images for learning content
    
  contact:
    name: Describe It Support
    email: support@describe-it.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://describe-it.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /descriptions/generate:
    post:
      tags:
        - Descriptions
      summary: Generate Image Descriptions
      description: |
        Generate AI-powered descriptions of images in multiple styles and both Spanish and English.
        Supports 5 different narrative styles for varied learning experiences.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDescriptionRequest'
            examples:
              narrativo:
                summary: Narrative style description
                value:
                  imageUrl: "https://images.unsplash.com/photo-example"
                  style: "narrativo"
                  maxLength: 300
              poetico:
                summary: Poetic style description
                value:
                  imageUrl: "https://images.unsplash.com/photo-example"
                  style: "poetico"
                  customPrompt: "Focus on emotions and metaphors"
                  maxLength: 250
      responses:
        '200':
          description: Descriptions generated successfully
          headers:
            X-Response-Time:
              description: API response time in milliseconds
              schema:
                type: string
            X-Fallback:
              description: Whether fallback content was used
              schema:
                type: string
                enum: ['true', 'false']
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DescriptionResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Descriptions
      summary: Description Service Info
      description: Get information about the description generation service
      responses:
        '200':
          description: Service information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceInfo'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GenerateDescriptionRequest:
      type: object
      required:
        - imageUrl
        - style
      properties:
        imageUrl:
          type: string
          format: uri
          description: URL of the image to describe
          example: "https://images.unsplash.com/photo-1234567890"
        style:
          type: string
          enum:
            - narrativo
            - poetico
            - academico
            - conversacional
            - infantil
          description: Narrative style for the description
          example: "narrativo"
        customPrompt:
          type: string
          description: Optional custom prompt for description generation
          example: "Focus on the emotional impact of the scene"
        maxLength:
          type: integer
          minimum: 50
          maximum: 1000
          default: 300
          description: Maximum length of the generated description
          example: 300
...
    DescriptionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Description'
        metadata:
          type: object
          properties:
            responseTime:
              type: string
              example: "1250.45ms"
            timestamp:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00Z"
            demoMode:
              type: boolean
              example: false
            fallback:
              type: boolean
              example: false
...
    ServiceInfo:
      type: object
      properties:
        service:
          type: string
          example: "Description Generation API"
        status:
          type: string
          example: "healthy"
        version:
          type: string
          example: "1.0.0"
        capabilities:
          type: object
          properties:
            styles:
              type: array
              items:
                type: string
              example: ["narrativo", "poetico", "academico"]
            languages:
              type: array
              items:
                type: string
              example: ["es", "en"]
            demoMode:
              type: boolean
              example: false

    # Entity Schemas
    Description:
      type: object
      properties:
        id:
          type: string
          example: "1705312200000_es"
        imageId:
          type: string
          description: Reference to the source image
          example: "unsplash_abc123"
        style:
          type: string
          enum: ["narrativo", "poetico", "academico", "conversacional", "infantil"]
          example: "narrativo"
        content:
          type: string
          description: The generated description text
          example: "Una hermosa puesta de sol se extiende sobre el océano infinito..."
        language:
          type: string
          enum: ["spanish", "english"]
          example: "spanish"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request parameters"
        code:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        details:
          type: object
          description: Additional error details (development only)
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid request parameters"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "imageUrl"
              message:
                type: string
                example: "Invalid image URL"
              code:
                type: string
                example: "invalid_url"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...